requirements.txt
python-slugify
discord
configparser